chore: inline transform function into webpack loader
