+++ packages/babel-plugin/src/index.tsx
@@ -1,3 +1,2 @@
 export { default } from './babel-plugin';
-export { transformAsync } from './transform';
 export type { PluginOptions, Metadata, State, Tag, TransformResult } from './types';
