+++ packages/webpack-loader/package.json
@@ -20,6 +20,7 @@
     "README.md"
   ],
   "dependencies": {
+    "@babel/core": "^7.13.10",
     "@compiled/babel-plugin": "0.6.7",
     "loader-utils": "2.0.0"
   },
