+++ packages/webpack-loader/src/__tests__/webpack-loader.test.tsx
@@ -1,7 +1,4 @@
 import loader from '../index';
-import { transformAsync } from '@compiled/babel-plugin';
-
-jest.mock('@compiled/babel-plugin');
 
 const getLoader = (opts: { addDependency: jest.Mock; callback: jest.Mock }) =>
   loader.bind({
@@ -12,65 +9,44 @@ const getLoader = (opts: { addDependency: jest.Mock; callback: jest.Mock }) =>
   });
 
 describe('webpack loader', () => {
-  it('should bail out early if the file has no compiled in it', () => {
-    const loader = getLoader({ addDependency: jest.fn(), callback: jest.fn() });
+  it('should bail out early if the file has no compiled in it', async () => {
+    const callback = jest.fn();
+    const loader = getLoader({ addDependency: jest.fn(), callback });
 
-    loader('console.log(undefined);');
+    await loader('console.log(undefined);');
 
-    expect(transformAsync).not.toHaveBeenCalled();
+    expect(callback).toHaveBeenCalled();
   });
 
-  it('should transform code if compiled has been found', () => {
-    const loader = getLoader({ addDependency: jest.fn(), callback: jest.fn() });
+  it('should transform code if compiled has been found', async () => {
+    const callback = jest.fn();
+    const loader = getLoader({ addDependency: jest.fn(), callback });
     const code = `import '@compiled/react';`;
 
-    loader(code);
+    await loader(code);
 
-    expect(transformAsync).toHaveBeenCalledWith(code, {
-      // Filename needed for module traversal
-      filename: '/projects/index.js',
-      opts: {
-        // Force caching module traversal calls
-        cache: true,
-        // Userland configuration
-        importReact: false,
-      },
-    });
+    expect(callback).toHaveBeenCalled();
   });
 
   it('should callback with transformed code', async () => {
     const callback = jest.fn();
-    (transformAsync as jest.Mock).mockReturnValue(
-      Promise.resolve({ code: 'transformed-code', includedFiles: [] })
-    );
     const loader = getLoader({ addDependency: jest.fn(), callback });
 
     await loader(`import '@compiled/react';`);
 
-    expect(callback).toHaveBeenCalledWith(null, 'transformed-code');
-  });
+    expect(callback.mock.calls[0][1]).toMatchInlineSnapshot(`
+      "/* index.js generated by @compiled/babel-plugin v0.0.0 */
 
-  it('should add dependencies for included files', async () => {
-    const addDependency = jest.fn();
-    (transformAsync as jest.Mock).mockReturnValue(
-      Promise.resolve({ code: 'transformed-code', includedFiles: ['one', 'two'] })
-    );
-    const loader = getLoader({ callback: jest.fn(), addDependency });
-
-    await loader(`import '@compiled/react';`);
-
-    expect(addDependency).toHaveBeenCalledWith('one');
-    expect(addDependency).toHaveBeenCalledWith('two');
-    expect(addDependency).toHaveBeenCalledTimes(2);
+      import { ax, ix, CC, CS } from \\"@compiled/react/runtime\\";"
+    `);
   });
 
   it('should callback on error', async () => {
     const callback = jest.fn();
-    (transformAsync as jest.Mock).mockReturnValue(Promise.reject('error occurred'));
     const loader = getLoader({ addDependency: jest.fn(), callback });
 
-    await loader(`import '@compiled/react';`);
+    await loader(`import '@compiled/react'; 222_!!!`);
 
-    expect(callback).toHaveBeenCalledWith('error occurred');
+    expect(callback.mock.calls[0][0]).toMatchSnapshot();
   });
 });
